requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.1
requests==2.32.3
Flask==3.1.0
cryptography==44.0.1
uwsgi==2.0.28